requirements.txt
-i https://pypi.org/simple
boto3==1.9.16
botocore==1.12.16
certifi==2018.8.24
chardet==3.0.4
docutils==0.14
idna==2.7
jmespath==0.9.3
python-dateutil==2.7.3; python_version >= '2.7'
pyyaml==3.13
requests==2.19.1
s3transfer==0.1.13
six==1.11.0
slackclient==1.3.0
urllib3==1.23
websocket-client==0.53.0
